apps/data-processing/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml==6.0.1
tenacity==8.2.3

# API clients
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
nitter-scraper==0.3.0

# Sentiment analysis
vaderSentiment==3.3.2
transformers>=4.11.0
tensorflow>=2.6.0
torch==2.1.1

# Storage
psycopg2-binary==2.9.9
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest>=6.2.5
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910

# Additional dependencies
scikit-learn>=0.24.2
tweepy>=4.0.0
python-dotenv>=0.19.0
ray>=1.9.0
